refactor: Make e2e test binary a folder structure
