+++ tests/e2e/main.rs
@@ -1,0 +1,3 @@
+mod account;
+mod common;
+mod token;
