+++ tests/e2e/token/mod.rs
@@ -1,8 +1,4 @@
-mod common;
-
-use common::setup_global;
 use hedera::{
-    AccountBalanceQuery,
     Client,
     TokenCreateTransaction,
     TokenDeleteTransaction,
@@ -17,6 +13,7 @@ use time::{
 use tokio::task::JoinSet;
 
 use crate::common::{
+    setup_global,
     Operator,
     TestEnvironment,
 };
@@ -121,23 +118,4 @@ async fn create_nft(client: &Client, token_id: TokenId) -> hedera::Result<Transa
         .metadata(Vec::from([Vec::from([])]))
         .execute(client)
         .await
-}
-
-#[tokio::test]
-async fn account_balance_query() -> anyhow::Result<()> {
-    let TestEnvironment { config, client } = setup_global();
-
-    let Some(op) = &config.operator else {
-        log::debug!("skipping test due to lack of operator");
-
-        return Ok(())
-    };
-
-    let balance = AccountBalanceQuery::new().account_id(op.account_id).execute(&client).await?;
-
-    log::trace!("successfully queried balance: {balance:?}");
-
-    anyhow::ensure!(balance.account_id == op.account_id);
-
-    Ok(())
 }
